Update FileStorage tests for relocated AzureBlobStorageService
